Add generic mechanism to make @Conditional match a specific k8s distro
