+++ pom.xml
@@ -77,7 +77,6 @@
 	<modules>
 		<module>spring-cloud-kubernetes-dependencies</module>
 		<module>spring-cloud-kubernetes-core</module>
-		<module>spring-cloud-openshift-core</module>
 		<module>spring-cloud-kubernetes-config</module>
 		<module>spring-cloud-kubernetes-discovery</module>
 		<module>spring-cloud-kubernetes-ribbon</module>
