+++ spring-cloud-kubernetes-core/src/test/java/org/springframework/cloud/kubernetes/condition/ServerUnreachableConditionSpringBootTest.java
@@ -10,6 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.cloud.kubernetes.condition.App.K8SBean;
+import org.springframework.cloud.kubernetes.condition.App.OpenshiftBean;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.junit4.SpringRunner;
 
@@ -24,6 +25,9 @@ public class ServerUnreachableConditionSpringBootTest {
 	@Autowired(required = false)
 	private K8SBean k8SBean;
 
+	@Autowired(required = false)
+	private OpenshiftBean openshiftBean;
+
 	@BeforeClass
 	public static void setUpBeforeClass() {
 
@@ -37,5 +41,10 @@ public class ServerUnreachableConditionSpringBootTest {
 	@Test
 	public void anyK8sBeanShouldNotBeCreated() {
 		assertNull(k8SBean);
+	}
+
+	@Test
+	public void openshiftBeanShouldNotBeCreated() {
+		assertNull(openshiftBean);
 	}
 }
