--- spring-cloud-openshift-core/src/main/java/org/springframework/cloud/kubernetes/condition/ConditionalOnOpenshift.java
@@ -1,16 +1,0 @@
-package org.springframework.cloud.kubernetes.condition;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-import org.springframework.context.annotation.Conditional;
-
-@Target({ ElementType.TYPE, ElementType.METHOD })
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-@Conditional(OnOpenshiftCondition.class)
-public @interface ConditionalOnOpenshift {
-
-}
